# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for منبر الجمعة (The Friday Pulpit)
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Text in all 6 supported languages
 */
export interface MultilingualText {
  ar: string;
  en: string;
  tr: string;
  fr: string;
  ur: string;
  fa: string;
}

export interface Khutbah {
  id: number;
  date: string;
  isCurrent: boolean;
  title: MultilingualText;
  body: MultilingualText;
  createdAt: string;
  updatedAt: string;
}

export interface KhutbahInput {
  date: string;
  isCurrent?: boolean;
  titleAr: string;
  titleEn: string;
  titleTr: string;
  titleFr: string;
  titleUr: string;
  titleFa: string;
  bodyAr: string;
  bodyEn: string;
  bodyTr: string;
  bodyFr: string;
  bodyUr: string;
  bodyFa: string;
}

export interface KhutbahUpdate {
  date?: string;
  isCurrent?: boolean;
  titleAr?: string;
  titleEn?: string;
  titleTr?: string;
  titleFr?: string;
  titleUr?: string;
  titleFa?: string;
  bodyAr?: string;
  bodyEn?: string;
  bodyTr?: string;
  bodyFr?: string;
  bodyUr?: string;
  bodyFa?: string;
}

export interface Settings {
  id: number;
  aboutAr: string;
  aboutEn: string;
  aboutTr: string;
  aboutFr: string;
  aboutUr: string;
  aboutFa: string;
}

export interface SettingsUpdate {
  aboutAr?: string;
  aboutEn?: string;
  aboutTr?: string;
  aboutFr?: string;
  aboutUr?: string;
  aboutFa?: string;
  currentPassword?: string;
  newPassword?: string;
}

export interface AdminPasswordInput {
  password: string;
}

export interface AdminVerifyResult {
  valid: boolean;
}

export interface AdminPasswordChange {
  currentPassword: string;
  newPassword: string;
}

export interface TranslationInput {
  titleAr: string;
  bodyAr: string;
  adminPassword: string;
}

export interface TranslationResult {
  titleEn: string;
  titleTr: string;
  titleFr: string;
  titleUr: string;
  titleFa: string;
  bodyEn: string;
  bodyTr: string;
  bodyFr: string;
  bodyUr: string;
  bodyFa: string;
}