# lib/api-zod/src/generated/types/adminPasswordChange.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for منبر الجمعة (The Friday Pulpit)
 * OpenAPI spec version: 0.1.0
 */

export interface AdminPasswordChange {
  currentPassword: string;
  newPassword: string;
}