# lib/api-zod/src/generated/types/khutbah.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for منبر الجمعة (The Friday Pulpit)
 * OpenAPI spec version: 0.1.0
 */
import type { MultilingualText } from "./multilingualText";

export interface Khutbah {
  id: number;
  date: Date;
  isCurrent: boolean;
  title: MultilingualText;
  body: MultilingualText;
  createdAt: Date;
  updatedAt: Date;
}