# lib/api-zod/src/generated/types/khutbahUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for منبر الجمعة (The Friday Pulpit)
 * OpenAPI spec version: 0.1.0
 */

export interface KhutbahUpdate {
  date?: Date;
  isCurrent?: boolean;
  titleAr?: string;
  titleEn?: string;
  titleTr?: string;
  titleFr?: string;
  titleUr?: string;
  titleFa?: string;
  bodyAr?: string;
  bodyEn?: string;
  bodyTr?: string;
  bodyFr?: string;
  bodyUr?: string;
  bodyFa?: string;
}