# lib/api-zod/src/generated/types/multilingualText.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for منبر الجمعة (The Friday Pulpit)
 * OpenAPI spec version: 0.1.0
 */

/**
 * Text in all 6 supported languages
 */
export interface MultilingualText {
  ar: string;
  en: string;
  tr: string;
  fr: string;
  ur: string;
  fa: string;
}