# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for منبر الجمعة (The Friday Pulpit)
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: number;
  aboutAr: string;
  aboutEn: string;
  aboutTr: string;
  aboutFr: string;
  aboutUr: string;
  aboutFa: string;
}