# lib/api-zod/src/generated/types/translationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for منبر الجمعة (The Friday Pulpit)
 * OpenAPI spec version: 0.1.0
 */

export interface TranslationInput {
  titleAr: string;
  bodyAr: string;
  adminPassword: string;
}