# lib/api-zod/src/generated/types/translationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for منبر الجمعة (The Friday Pulpit)
 * OpenAPI spec version: 0.1.0
 */

export interface TranslationResult {
  titleEn: string;
  titleTr: string;
  titleFr: string;
  titleUr: string;
  titleFa: string;
  bodyEn: string;
  bodyTr: string;
  bodyFr: string;
  bodyUr: string;
  bodyFa: string;
}